chore(refactor-codebase): refactor the codebase
